requirements.txt
Django==3.2
django-cors-headers==3.12.0
django-extensions==3.2.0
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
djangorestframework-simplejwt==5.2.0
